Refactor to object props for recipe reads and fix fs-extra open
