+++ websites/portfolio/editor/src/app/image/[...filePath]/route.ts
@@ -1,6 +1,6 @@
 import { getContentDirectory } from "content-engine/fs/getContentDirectory";
 import { ReadStream } from "fs";
-import { open } from "fs-extra";
+import { open } from "fs/promises";
 import { notFound } from "next/navigation";
 import { NextRequest, NextResponse } from "next/server";
 import { join, resolve } from "path";
