+++ websites/portfolio/editor/src/app/uploads/[filename]/route.ts
@@ -1,6 +1,6 @@
 import { uploadsDirectory } from "portfolio-website-common/homepage-controller/paths";
 import { ReadStream } from "fs";
-import { open } from "fs-extra";
+import { open } from "fs/promises";
 import { notFound } from "next/navigation";
 import { NextRequest, NextResponse } from "next/server";
 import { resolve } from "path";
