+++ websites/recipe-website/editor/src/app/(recipes)/recipe/[slug]/copy/page.tsx
@@ -19,7 +19,7 @@ export default async function Recipe({
   }
   let recipe;
   try {
-    recipe = await getRecipeBySlug(slug);
+    recipe = await getRecipeBySlug({ slug });
   } catch (e) {
     if (e instanceof Error && "code" in e && e.code === "ENOENT") {
       notFound();
