+++ websites/recipe-website/editor/src/app/(recipes)/recipe/[slug]/edit/page.tsx
@@ -20,7 +20,7 @@ export default async function Recipe({
   }
   let recipe;
   try {
-    recipe = await getRecipeBySlug(slug);
+    recipe = await getRecipeBySlug({ slug });
   } catch (e) {
     if (e instanceof Error && "code" in e && e.code === "ENOENT") {
       notFound();
