+++ websites/recipe-website/editor/src/app/uploads/[filename]/route.ts
@@ -1,6 +1,6 @@
 import { getContentDirectory } from "content-engine/fs/getContentDirectory";
 import { ReadStream } from "fs";
-import { open } from "fs-extra";
+import { open } from "fs/promises";
 import { notFound } from "next/navigation";
 import { NextRequest, NextResponse } from "next/server";
 import { resolve } from "path";
