+++ websites/recipe-website/export/src/app/(recipes)/recipe/[slug]/page.tsx
@@ -14,7 +14,7 @@ export async function generateMetadata({
   const { slug } = await params;
   let recipe;
   try {
-    recipe = await getCachedRecipeBySlug(slug);
+    recipe = await getCachedRecipeBySlug({ slug });
   } catch (e) {
     if (e instanceof Error && "code" in e && e.code === "ENOENT") {
       notFound();
@@ -32,7 +32,7 @@ export default async function Recipe({
   const { slug } = await params;
   let recipe;
   try {
-    recipe = await getRecipeBySlug(slug);
+    recipe = await getRecipeBySlug({ slug });
   } catch (e) {
     if (e instanceof Error && "code" in e && e.code === "ENOENT") {
       notFound();
